Improve main/menu_cmd matching
CmdDraw canUse EquipChk arm: cast EquipChk's int return to u8 before the
!= 0 test (static_cast<u8>(EquipChk(...)) != 0). The debug target tests
EquipChk's low byte (clrlwi r3,r3,24) before the bool idiom; the bare int
!= 0 compared the full word. 68.30 -> 68.61, unit 74.921 -> 74.996.

+++ src/menu_cmd.cpp
@@ -1146,7 +1146,7 @@ void CMenuPcs::CmdDraw()
 								canUse = static_cast<u32>(
 								    ChkUnite(GetCmdStateView(this)->selected, combo) > 0);
 							} else {
-								canUse = static_cast<u32>(EquipChk(static_cast<int>(canBuf[sel - 1])) != 0);
+								canUse = static_cast<u32>(static_cast<u8>(EquipChk(static_cast<int>(canBuf[sel - 1]))) != 0);
 							}
 
 							if (canUse == 0) {
